Rename graph_conv_predictor
